Keep polygons that do not intersect stream network
The code was getting rid of polygons that do not intersect stream network but I think they should be kept, because this mostly occurs when the stream network is off and the polygons are in fact just fine.

PS: I kept the lines and introduced a boolean to allow for easily reactivating this processing if one wants it.